+++ app/src/components/FloatingPlusButton.tsx
@@ -1,12 +1,16 @@
 import React from "react";
 import { StyleSheet } from "react-native";
 import Animated, { useAnimatedStyle } from "react-native-reanimated";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+
 import { colors } from "../utils/colors";
 import Button from "./RoundButtonIcon";
 import logEvents from "../services/logEvents";
+import { TAB_BAR_HEIGHT } from "@/utils/constants";
 
 const FloatingPlusButton = ({ onPress, shadow, plusPosition }) => {
   // Handle both regular numbers and SharedValues
+  const insets = useSafeAreaInsets();
   const animatedStyle = useAnimatedStyle(() => {
     const translateX = typeof plusPosition === "object" && plusPosition?.value !== undefined ? plusPosition.value : plusPosition || 0;
 
@@ -18,7 +22,15 @@ const FloatingPlusButton = ({ onPress, shadow, plusPosition }) => {
   if (!plusPosition && plusPosition !== 0) return null;
 
   return (
-    <Animated.View style={[styles.buttonWrapper, animatedStyle]}>
+    <Animated.View
+      style={[
+        styles.buttonWrapper,
+        {
+          bottom: insets.bottom + TAB_BAR_HEIGHT,
+        },
+        animatedStyle,
+      ]}
+    >
       <Button
         backgroundColor={colors.LIGHT_BLUE}
         iconColor={colors.WHITE}