+++ app/src/scenes/suivi/events/index.tsx
@@ -1,21 +1,21 @@
 import React from "react";
-import { StyleSheet, View, ScrollView } from "react-native";
+import { StyleSheet, View, ScrollView, Text, Dimensions } from "react-native";
 import { useFocusEffect } from "@react-navigation/native";
 import { isToday, isYesterday, parseISO } from "date-fns";
-import { getArrayOfDatesFromTo, formatDay, formatRelativeDate } from "../../../utils/date/helpers";
-import { DiaryDataContext } from "../../../context/diaryData";
-import Text from "../../../components/MyText";
-import { colors } from "../../../utils/colors";
-import { buildSurveyData } from "../../survey/survey-data";
-import Icon from "../../../components/Icon";
-import localStorage from "../../../utils/localStorage";
-import logEvents from "../../../services/logEvents";
-import Button from "../../../components/Button";
+import { getArrayOfDatesFromTo, formatDay, formatRelativeDate } from "@/utils/date/helpers";
+import { DiaryDataContext } from "@/context/diaryData";
+import { colors } from "@/utils/colors";
+import Icon from "@/components/Icon";
+import localStorage from "@/utils/localStorage";
+import logEvents from "@/services/logEvents";
 import Card from "./Card";
 import { EventFilterHeader } from "./EventFilterHeader";
 import JMButton from "@/components/JMButton";
-import Legend from "../Legend";
 import { getIndicatorKey } from "@/utils/indicatorUtils";
+import { TAB_BAR_HEIGHT } from "@/utils/constants";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+
+const screenHeight = Dimensions.get("window").height;
 
 const Events = ({ navigation, presetDate, setPresetDate, fromDate, setFromDate, toDate, setToDate, onScroll }) => {
   const [diaryData] = React.useContext(DiaryDataContext);
@@ -27,6 +27,7 @@ const Events = ({ navigation, presetDate, setPresetDate, fromDate, setFromDate, 
   const [indicateurId, setIndicateurId] = React.useState();
   const [event, setEvent] = React.useState("ALL");
   const [level, setLevel] = React.useState([5]);
+  const insets = useSafeAreaInsets();
 
   useFocusEffect(
     React.useCallback(() => {
@@ -155,7 +156,18 @@ const Events = ({ navigation, presetDate, setPresetDate, fromDate, setFromDate, 
   }
   return (
     <>
-      <ScrollView style={styles.scrollView} contentContainerStyle={styles.scrollContainer} onScroll={onScroll}>
+      <ScrollView
+        style={styles.scrollView}
+        contentContainerStyle={[
+          styles.scrollContainer,
+          {
+            paddingBottom: insets.bottom + TAB_BAR_HEIGHT,
+          },
+        ]}
+        showsVerticalScrollIndicator={false}
+        scrollEventThrottle={16}
+        onScroll={onScroll}
+      >
         <EventFilterHeader
           presetDate={presetDate}
           setPresetDate={setPresetDate}
@@ -164,7 +176,11 @@ const Events = ({ navigation, presetDate, setPresetDate, fromDate, setFromDate, 
           toDate={toDate}
           setToDate={setToDate}
           indicateur={indicateur}
-          setIndicateur={setIndicateur}
+          setIndicateur={(indicatorName) => {
+            setIndicateur(indicatorName);
+            const _indicator = userIndicateurs.find((ind) => ind.name === indicatorName);
+            setIndicateurId(getIndicatorKey(_indicator));
+          }}
           level={level}
           setLevel={setLevel}
           userIndicateurs={userIndicateurs.filter(({ active }) => active)}
@@ -271,7 +287,7 @@ const styles = StyleSheet.create({
   },
   scrollContainer: {
     paddingTop: 20,
-    paddingBottom: 30,
+    minHeight: screenHeight * 0.7,
   },
   dataContainer: {
     marginTop: 40,